Change html-related folder structure
