--- lib/inquisition/formatters/html_formatter.rb
@@ -1,53 +1,0 @@
-require 'fileutils'
-require 'inquisition/formatters/html/builder'
-require 'inquisition/formatters/html/overview_builder'
-require 'inquisition/formatters/html/issues_list_builder'
-
-module Inquisition
-  module Formatters
-    class HtmlFormatter
-      def initialize(collection)
-        @collection = collection
-      end
-
-      def call
-        create_directories
-        copy_assets
-        create_files
-      end
-
-      private
-
-      def create_directories
-        FileUtils.mkdir_p(report_directory)
-      end
-
-      def copy_assets
-        FileUtils.cp_r(assets_directory, report_directory)
-      end
-
-      def create_files
-        generators.each do |generator|
-          File.open(generator.file_path, 'w+') do |file|
-            file.write(generator.render)
-          end
-        end
-      end
-
-      def generators
-        [
-          HTML::IssuesListBuilder.new(@collection),
-          HTML::OverviewBuilder.new(@collection)
-        ]
-      end
-
-      def assets_directory
-        File.join(Inquisition.root, 'lib/inquisition/formatters/html/assets')
-      end
-
-      def report_directory
-        File.join(Rails.root, 'inquisition')
-      end
-    end
-  end
-end
