+++ lib/inquisition/outputter/html_output/overview/breakdowns_builder.rb
@@ -1,6 +1,6 @@
 module Inquisition
-  module Formatters
-    module HTML
+  module Outputter
+    module HtmlOutput
       module Overview
         class BreakdownsBuilder < Builder
           def file_name
