+++ lib/inquisition/outputter/html_output/overview/coverage_builder.rb
@@ -1,6 +1,6 @@
 module Inquisition
-  module Formatters
-    module HTML
+  module Outputter
+    module HtmlOutput
       module Overview
         class CoverageBuilder < Builder
           def file_name
