+++ views/overview.html.erb
@@ -9,14 +9,14 @@
   <meta name="description" content="">
   <meta name="author" content="">
 
-  <title>Inquisition: issues</title>
+  <title>Inquisition: overview</title>
 
-  <!-- Custom fonts for this template -->
-  <link href="assets/vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
+  <!-- Custom fonts for this template-->
+  <link href="assets/css/vendor/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
   <link href="https://fonts.googleapis.com/css?family=Nunito:200,200i,300,300i,400,400i,600,600i,700,700i,800,800i,900,900i" rel="stylesheet">
 
-  <!-- Custom styles for this template -->
-  <link href="assets/css/sb-admin-2.min.css" rel="stylesheet">
+  <!-- Custom styles for this template-->
+  <link href="assets/css/vendor/sb-admin-2.min.css" rel="stylesheet">
 
 </head>
 
@@ -72,8 +72,7 @@
 
         <!-- Topbar -->
         <nav class="navbar navbar-expand navbar-light bg-white topbar mb-4 static-top shadow">
-
-
+          <!-- Topbar content -->
         </nav>
         <!-- End of Topbar -->
 
@@ -81,33 +80,22 @@
         <div class="container-fluid">
 
           <!-- Page Heading -->
-          <h1 class="h3 mb-2 text-gray-800">Issues</h1>
-          <p class="mb-4">Some description</p>
+          <div class="d-sm-flex align-items-center justify-content-between mb-4">
+            <h1 class="h3 mb-0 text-gray-800">Overview</h1>
+          </div>
 
-          <!-- DataTales Example -->
-          <% @collection.each_with_index do |issue, index| %>
-            <div class="card shadow mb-4">
-              <div class="card-header py-3">
-                <h6 class="m-0 font-weight-bold text-primary"># <%= index + 1 %></h6>
-              </div>
-              <div class="card-body">
-                <p><%= issue.message %></p>
-                <% if issue.path %>
-                  <span class="small">
-                    <span>Found in </span>
-                    <span class="text-success"><%= issue.path %>, line <%= issue.line %></span>
-                    <span> by </span>
-                    <span class="text-success"><%= issue.runner.class.badge %></span>
-                  </span>
-                <% else %>
-                  <span class="small">
-                    <span>Found by </span>
-                    <span class="text-success"><%= issue.runner.class.badge %></span>
-                  </span>
-                <% end %>
-              </div>
-            </div>
-          <% end %>
+          <div class="row">
+
+            <!-- Area Chart -->
+
+            <%= main_field.render %>
+
+            <!-- Pie Chart -->
+
+            <%= breakdowns.render %>
+
+          </div>
+
         </div>
         <!-- /.container-fluid -->
 
@@ -135,17 +123,17 @@
     <i class="fas fa-angle-up"></i>
   </a>
 
-
   <!-- Bootstrap core JavaScript-->
-  <script src="assets/vendor/jquery/jquery.min.js"></script>
-  <script src="assets/vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
+  <script src="assets/js/vendor/jquery.min.js"></script>
+  <script src="assets/js/vendor/bootstrap.bundle.min.js"></script>
 
   <!-- Core plugin JavaScript-->
-  <script src="assets/vendor/jquery-easing/jquery.easing.min.js"></script>
+  <script src="assets/js/vendor/jquery.easing.min.js"></script>
 
   <!-- Custom scripts for all pages-->
-  <script src="assets/js/sb-admin-2.min.js"></script>
-
+  <script src="assets/js/vendor/sb-admin-2.min.js"></script>
+  <script src="assets/js/vendor/Chart.min.js"></script>
+  <script src="assets/js/breakdowns-chart.js"></script>
 </body>
 
 </html>
